Loosen the restriction on cursor character set
As it stands ("limited to the base64 character set"), there is some confusion as to exactly _which_ base64 variant is allowed. The "standard" base64 variant is what one usually mean when just saying "base64" without qualification, however the URL-safe base64 variant is more appropriate for use in URLs. Both variants are defined in RFC4648.

We could modify the sentence to specify which variant(s) of base64 are allowed, but it seems better to just relax this restriction a bit, as the intention isn't to force implementors to encode the cursor in a specific way, but rather to impose some minimal restrictions on what character values the cursor can have, so that one doesn't have to consider which character set the cursor is encoded in, or handle raw byte values.